changed storage type from uint8_t to double

+++ include/data.hpp
@@ -4,19 +4,19 @@
 #include <vector>
 
 class Data {
-  std::vector<uint8_t> *feature_vector_;
+  std::vector<double> *feature_vector_;
   uint8_t label_ {};
 public:
   Data();
   ~Data();
   explicit Data(size_t);
-  void set_feature_vector(std::vector<uint8_t>*);
-  void append_to_feature_vector(uint8_t) const;
+  void set_feature_vector(std::vector<double>*);
+  void append_to_feature_vector(double) const;
   void set_label(uint8_t);
 
   [[nodiscard]] size_t get_feature_vector_size() const;
   [[nodiscard]] uint8_t get_label() const;
-  [[nodiscard]] std::vector<uint8_t>* get_feature_vector() const;
+  [[nodiscard]] std::vector<double>* get_feature_vector() const;
 };
 
 #endif //DATA_HPP
